requirements.txt
pandas~=2.2.3
matplotlib~=3.10.0
seaborn~=0.13.2
scikit-learn~=1.6.1
numpy~=2.2.2

scipy~=1.15.1
torch_geometric
